requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv

langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters

langchain-postgres
psycopg[binary]
sqlalchemy
alembic
pypdf
python-docx
cryptography>=3.1
bcrypt
PyJWT
tqdm

sentence-transformers

pytest
httpx
